refactor: move period table functions to table-display module
